packages/interfaces: create src/session.ts
