+++ packages/interfaces/src/index.ts
@@ -3,3 +3,4 @@ export * from "./signer.js"
 export * from "./values.js"
 export * from "./action.js"
 export * from "./message.js"
+export * from "./session.js"
